backend/rag-chatbot/requirements.txt
# RAG Agent Service Requirements

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Vector database
qdrant-client==1.7.0

# OpenAI SDK (for Agents SDK compatibility)
openai==1.3.7

# Google Generative AI (for Gemini)
google-generativeai==0.4.1

# Configuration and environment
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
structlog==23.1.0

# Utilities
tqdm==4.66.1
python-slugify==8.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Async utilities
asyncio
aiofiles==23.2.1